+++ drv/local-vpd/Cargo.toml
@@ -4,15 +4,12 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-tlvc = { workspace = true }
 zerocopy = { workspace = true }
 
-drv-i2c-api = { path = "../../drv/i2c-api" }
+drv-oxide-vpd = { path = "../oxide-vpd" }
 drv-i2c-devices = { path = "../../drv/i2c-devices" }
+drv-i2c-api = { path = "../../drv/i2c-api" }
 userlib = { path = "../../sys/userlib" }
-ringbuf = { path = "../../lib/ringbuf" }
 
 [build-dependencies]
-serde = { workspace = true }
 build-i2c = {path = "../../build/i2c" }
-build-util = {path = "../../build/util"}
