+++ drv/local-vpd/src/lib.rs
@@ -9,65 +9,11 @@
 //! `read_config` reads from the *local* EEPROM; i.e. is the one soldered to the
 //! PCB itself.  The app TOML file must have one AT24xx named `local_vpd`; we
 //! use that name to pick which EEPROM to read in `read_config`.
-//!
-//! The system may have additional EEPROMs on FRUs that plug into the board
-//! (e.g. fans); those can be read with `read_config_from`. We assume that the
-//! all EEPROMs are AT24CSW080s, and that they contains keys in TLV-C format
-//! (see RFD 148 for a general description, or RFD 320 for the specific example
-//! of MAC addresses)
 
-use drv_i2c_devices::at24csw080::{
-    At24Csw080, Error as At24Error, EEPROM_SIZE,
-};
-use ringbuf::*;
-use tlvc::{TlvcRead, TlvcReadError, TlvcReader};
-use userlib::*;
+use drv_i2c_devices::at24csw080::At24Csw080;
+use drv_oxide_vpd::VpdError;
+use userlib::TaskId;
 use zerocopy::{AsBytes, FromBytes};
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum LocalVpdError {
-    ErrorOnBegin(TlvcReadError<At24Error>),
-    ErrorOnRead(TlvcReadError<At24Error>),
-    ErrorOnNext(TlvcReadError<At24Error>),
-    NoSuchChunk([u8; 4]),
-    InvalidChecksum(TlvcReadError<At24Error>),
-    InvalidChunkSize,
-    /// The base FRU0 chunk we expected was not present.
-    NoRootChunk,
-}
-
-#[derive(Clone)]
-struct EepromReader<'a> {
-    eeprom: &'a At24Csw080,
-}
-
-#[derive(Copy, Clone, PartialEq)]
-enum Trace {
-    EepromError(drv_i2c_devices::at24csw080::Error),
-    Error(LocalVpdError),
-    UnrelatedChunk([u8; 4]),
-    None,
-}
-
-ringbuf!(Trace, 4, Trace::None);
-
-impl<'a> TlvcRead for EepromReader<'a> {
-    type Error = drv_i2c_devices::at24csw080::Error;
-    fn extent(&self) -> Result<u64, TlvcReadError<Self::Error>> {
-        Ok(EEPROM_SIZE as u64)
-    }
-    fn read_exact(
-        &self,
-        offset: u64,
-        dest: &mut [u8],
-    ) -> Result<(), TlvcReadError<Self::Error>> {
-        self.eeprom.read_into(offset as u16, dest).map_err(|code| {
-            ringbuf_entry!(Trace::EepromError(code));
-            TlvcReadError::User(code)
-        })?;
-        Ok(())
-    }
-}
 
 /// Searches for the given TLV-C tag in the local VPD and reads it
 ///
@@ -90,60 +36,10 @@ impl<'a> TlvcRead for EepromReader<'a> {
 pub fn read_config<V: AsBytes + FromBytes>(
     i2c_task: TaskId,
     tag: [u8; 4],
-) -> Result<V, LocalVpdError> {
-    let eeprom = drv_i2c_devices::at24csw080::At24Csw080::new(
-        i2c_config::devices::at24csw080_local_vpd(i2c_task),
-    );
-    read_config_from(eeprom, tag)
-}
-
-pub fn read_config_from<V: AsBytes + FromBytes>(
-    eeprom: At24Csw080,
-    tag: [u8; 4],
-) -> Result<V, LocalVpdError> {
-    let mut out = V::new_zeroed();
-    let n = read_config_from_into(eeprom, tag, out.as_bytes_mut())?;
-
-    // `read_config_into()` fails if the data is too large for `out`, but will
-    // succeed if it's less than out; we want to guarantee it's exactly the size
-    // of V.
-    if n != core::mem::size_of::<V>() {
-        return Err(LocalVpdError::InvalidChunkSize);
-    }
-
-    Ok(out)
-}
-
-/// Searches for the given TLV-C tag in the given VPD and reads it
-///
-/// Returns an error if the tag is not present, the data is too large to fit in
-/// `out`, or any checksum is corrupt.
-///
-/// The data in the EEPROM is assumed to be of the form
-/// ```ron
-/// ("FRU0", [
-///     ("TAG1", [ [...] ]),
-///     ("TAG2", [ [...] ]),
-///     ("TAG3", [ [...] ]),
-/// ])
-/// ```
-/// (where `TAG*` are example tags)
-///
-/// `read_config` should be called with a tag nested under `FRU0` (e.g. `TAG1`
-/// in the example above).  It will copy the raw byte array (shown as
-/// `[...]`) into `out`, returning the number of bytes written.
-pub fn read_config_from_into(
-    eeprom: At24Csw080,
-    tag: [u8; 4],
-    out: &mut [u8],
-) -> Result<usize, LocalVpdError> {
-    match read_config_inner(eeprom, tag, out) {
-        Ok(n) => Ok(n),
-        Err(e) => {
-            ringbuf_entry!(Trace::Error(e));
-            Err(e)
-        }
-    }
+) -> Result<V, VpdError> {
+    let eeprom =
+        At24Csw080::new(i2c_config::devices::at24csw080_local_vpd(i2c_task));
+    drv_oxide_vpd::read_config_from(eeprom, tag)
 }
 
 /// Searches for the given TLV-C tag in the local VPD and reads it
@@ -153,62 +49,10 @@ pub fn read_config_into(
     i2c_task: TaskId,
     tag: [u8; 4],
     out: &mut [u8],
-) -> Result<usize, LocalVpdError> {
-    let eeprom = drv_i2c_devices::at24csw080::At24Csw080::new(
-        i2c_config::devices::at24csw080_local_vpd(i2c_task),
-    );
-    read_config_from_into(eeprom, tag, out)
-}
-
-/// Implementation factor of `read_config_into` above to ensure that all errors
-/// are recorded. Any error returned from this routine will be put into a
-/// ringbuf by its caller, so it needn't worry about it.
-fn read_config_inner(
-    eeprom: At24Csw080,
-    tag: [u8; 4],
-    out: &mut [u8],
-) -> Result<usize, LocalVpdError> {
-    let eeprom_reader = EepromReader { eeprom: &eeprom };
-
-    let mut reader = TlvcReader::begin(eeprom_reader)
-        .map_err(LocalVpdError::ErrorOnBegin)?;
-
-    loop {
-        match reader.next() {
-            Ok(Some(chunk)) => {
-                let mut scratch = [0u8; 32];
-                if chunk.header().tag == *b"FRU0" {
-                    chunk
-                        .check_body_checksum(&mut scratch)
-                        .map_err(LocalVpdError::InvalidChecksum)?;
-                    let mut inner = chunk.read_as_chunks();
-                    while let Ok(Some(chunk)) = inner.next() {
-                        if chunk.header().tag == tag {
-                            chunk
-                                .check_body_checksum(&mut scratch)
-                                .map_err(LocalVpdError::InvalidChecksum)?;
-
-                            let chunk_len = chunk.len() as usize;
-
-                            if chunk_len > out.len() {
-                                return Err(LocalVpdError::InvalidChunkSize);
-                            }
-
-                            chunk
-                                .read_exact(0, &mut out[..chunk_len])
-                                .map_err(LocalVpdError::ErrorOnRead)?;
-                            return Ok(chunk_len);
-                        }
-                    }
-                    return Err(LocalVpdError::NoSuchChunk(tag));
-                } else {
-                    ringbuf_entry!(Trace::UnrelatedChunk(chunk.header().tag));
-                }
-            }
-            Ok(None) => return Err(LocalVpdError::NoRootChunk),
-            Err(e) => return Err(LocalVpdError::ErrorOnNext(e)),
-        }
-    }
+) -> Result<usize, VpdError> {
+    let eeprom =
+        At24Csw080::new(i2c_config::devices::at24csw080_local_vpd(i2c_task));
+    drv_oxide_vpd::read_config_from_into(eeprom, tag, out)
 }
 
 include!(concat!(env!("OUT_DIR"), "/i2c_config.rs"));
