+++ task/host-sp-comms/Cargo.toml
@@ -20,7 +20,7 @@ drv-gimlet-hf-api.path= "../../drv/gimlet-hf-api"
 drv-gimlet-seq-api.path= "../../drv/gimlet-seq-api"
 drv-i2c-api.path= "../../drv/i2c-api"
 drv-i2c-devices.path= "../../drv/i2c-devices"
-drv-local-vpd.path= "../../drv/local-vpd"
+drv-oxide-vpd.path= "../../drv/oxide-vpd"
 drv-stm32xx-sys-api.path= "../../drv/stm32xx-sys-api"
 host-sp-messages.path= "../../lib/host-sp-messages"
 multitimer.path= "../../lib/multitimer"
