+++ task/host-sp-comms/src/bsp.rs
@@ -1,0 +1,6 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+// This file should never be included; if it is, the build configuration is bad
+compile_error!("no BSP for the given target board");
